Ticket:User-Facing Backend
users will be able to register for a new account, log in, logout, and delete their account.

Registering should create an account and log the user in, ensuring an entry is made in the sessions collection. There is a unique index on the user_id field in sessions, so we can efficiently query on this field