Ticket:Durable Writes
When a new user registers for MFlix, their information must be added to the database before they can do anything else on the site. For this reason, we want to make sure that the data written by the add_user method will not be rolled back.

We can completely eliminate the chances of a rollback by increasing the write durability of the add_user method.